# Quiz-Genie/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Quivio - AI-Assisted Quiz Generation Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /questions/{id}:
    patch:
      operationId: updateQuestion
      tags: [questions]
      summary: Update a question (text, options, answer)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuestionBody"
      responses:
        "200":
          description: Updated question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
    delete:
      operationId: deleteQuestion
      tags: [questions]
      summary: Delete a question
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /attempts:
    get:
      operationId: listAttempts
      tags: [attempts]
      summary: List quiz attempts for the current user
      responses:
        "200":
          description: List of attempts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attempt"
    post:
      operationId: createAttempt
      tags: [attempts]
      summary: Start a new quiz attempt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAttemptBody"
      responses:
        "201":
          description: Created attempt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attempt"
  /attempts/{id}:
    get:
      operationId: getAttempt
      tags: [attempts]
      summary: Get a specific attempt with answers
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Attempt with answers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttemptWithAnswers"
  /attempts/{id}/submit:
    post:
      operationId: submitAttempt
      tags: [attempts]
      summary: Submit answers and get AI-evaluated results
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitAttemptBody"
      responses:
        "200":
          description: Attempt result with AI feedback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttemptResult"
...
  /dashboard/quiz-performance/{quizId}:
    get:
      operationId: getQuizPerformance
      tags: [dashboard]
      summary: Get performance analytics for a specific quiz (faculty)
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quiz performance analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizPerformance"
components:
  schemas:
...
    Question:
      type: object
      properties:
        id:
          type: integer
        quizId:
          type: integer
        questionText:
          type: string
        questionType:
          type: string
          enum: [multiple_choice, short_answer]
        options:
          type: ["string", "null"]
          description: JSON array of options for multiple choice
        correctAnswer:
          type: string
        explanation:
          type: string
        orderIndex:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - quizId
        - questionText
        - questionType
        - correctAnswer
        - explanation
        - orderIndex
        - createdAt
    UpdateQuestionBody:
      type: object
      properties:
        questionText:
          type: string
        options:
          type: ["string", "null"]
        correctAnswer:
          type: string
        explanation:
          type: string
    Attempt:
      type: object
      properties:
        id:
          type: integer
        quizId:
          type: integer
        userId:
          type: string
        score:
          type: ["number", "null"]
        maxScore:
          type: ["integer", "null"]
        percentage:
          type: ["number", "null"]
        status:
          type: string
          enum: [in_progress, completed]
        feedback:
          type: ["string", "null"]
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: ["string", "null"]
        quizTitle:
          type: ["string", "null"]
      required:
        - id
        - quizId
        - userId
        - status
        - startedAt
    AttemptWithAnswers:
      type: object
      properties:
        id:
          type: integer
        quizId:
          type: integer
        userId:
          type: string
        score:
          type: ["number", "null"]
        maxScore:
          type: ["integer", "null"]
        percentage:
          type: ["number", "null"]
        status:
          type: string
          enum: [in_progress, completed]
        feedback:
          type: ["string", "null"]
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: ["string", "null"]
        quizTitle:
          type: ["string", "null"]
        answers:
          type: array
          items:
            $ref: "#/components/schemas/Answer"
      required:
        - id
        - quizId
        - userId
        - status
        - startedAt
        - answers
    CreateAttemptBody:
      type: object
      properties:
        quizId:
          type: integer
      required:
        - quizId
    SubmitAttemptBody:
      type: object
      properties:
        answers:
          type: array
          items:
            type: object
            properties:
              questionId:
                type: integer
              userAnswer:
                type: string
            required:
              - questionId
              - userAnswer
      required:
        - answers
    AttemptResult:
      type: object
      properties:
        id:
          type: integer
        quizId:
          type: integer
        userId:
          type: string
        score:
          type: number
        maxScore:
          type: integer
        percentage:
          type: number
        status:
          type: string
          enum: [completed]
        feedback:
          type: string
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        quizTitle:
          type: ["string", "null"]
        answers:
          type: array
          items:
            $ref: "#/components/schemas/Answer"
      required:
        - id
        - quizId
        - userId
        - score
        - maxScore
        - percentage
        - status
        - feedback
        - startedAt
        - completedAt
        - answers
    Answer:
      type: object
      properties:
        id:
          type: integer
        attemptId:
          type: integer
        questionId:
          type: integer
        userAnswer:
          type: string
        isCorrect:
          type: ["boolean", "null"]
        pointsEarned:
          type: ["number", "null"]
        aiFeedback:
          type: ["string", "null"]
        questionText:
          type: ["string", "null"]
        correctAnswer:
          type: ["string", "null"]
        explanation:
          type: ["string", "null"]
      required:
        - id
        - attemptId
        - questionId
        - userAnswer
...
    QuizPerformance:
      type: object
      properties:
        quizId:
          type: integer
        quizTitle:
          type: string
        totalAttempts:
          type: integer
        averageScore:
          type: number
        averagePercentage:
          type: number
        passRate:
          type: number
        questionStats:
          type: array
          items:
            $ref: "#/components/schemas/QuestionStat"
      required:
        - quizId
        - quizTitle
        - totalAttempts
        - averageScore
        - averagePercentage
        - passRate
        - questionStats
    QuestionStat:
      type: object
      properties:
        questionId:
          type: integer
        questionText:
          type: string
        correctRate:
          type: number
        totalAnswers:
          type: integer
      required:
        - questionId
        - questionText
        - correctRate
        - totalAnswers