# Quiz-Genie/lib/api-zod/src/generated/types/attempt.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quivio - AI-Assisted Quiz Generation Platform
 * OpenAPI spec version: 0.1.0
 */
import type { AttemptStatus } from "./attemptStatus";

export interface Attempt {
  id: number;
  quizId: number;
  userId: string;
  /** @nullable */
  score?: number | null;
  /** @nullable */
  maxScore?: number | null;
  /** @nullable */
  percentage?: number | null;
  status: AttemptStatus;
  /** @nullable */
  feedback?: string | null;
  startedAt: Date;
  /** @nullable */
  completedAt?: string | null;
  /** @nullable */
  quizTitle?: string | null;
}