# Quiz-Genie/lib/api-zod/src/generated/types/attemptResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quivio - AI-Assisted Quiz Generation Platform
 * OpenAPI spec version: 0.1.0
 */
import type { Answer } from "./answer";
import type { AttemptResultStatus } from "./attemptResultStatus";

export interface AttemptResult {
  id: number;
  quizId: number;
  userId: string;
  score: number;
  maxScore: number;
  percentage: number;
  status: AttemptResultStatus;
  feedback: string;
  startedAt: Date;
  completedAt: Date;
  /** @nullable */
  quizTitle?: string | null;
  answers: Answer[];
}