# Quiz-Genie/lib/api-zod/src/generated/types/submitAttemptBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quivio - AI-Assisted Quiz Generation Platform
 * OpenAPI spec version: 0.1.0
 */
import type { SubmitAttemptBodyAnswersItem } from "./submitAttemptBodyAnswersItem";

export interface SubmitAttemptBody {
  answers: SubmitAttemptBodyAnswersItem[];
}